chore : 테스트 전이 Lottostest -> LottoGeneratorTest
